pyproject.toml
[project]
name = "voiye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.127.0",
    "ipywidgets>=8.1.8",
    "openai>=2.14.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "supabase>=2.27.0",
    "types-pyyaml>=6.0.12.20250915",
    "ultralytics>=8.3.240",
    "uvicorn>=0.40.0",
]
